Criei arquivos para organizar métodos repetidos.
