+++ semana-10/projeto/labex/src/assets/constants/axiosConstants.js
@@ -1,0 +1,2 @@
+export const baseUrl = 
+  "https://us-central1-labenu-apis.cloudfunctions.net/labeX/roberto-salgado-jackson"
