+++ semana-10/projeto/labex/src/assets/hooks/useInput.js
@@ -1,0 +1,11 @@
+import { useState } from 'react'
+
+export const useInput = () => {
+  const [value, updateValue] = useState ("")
+
+  const handleValue = event => {
+    updateValue(event.target.value)
+  }
+
+  return [value, handleValue]
+}
