+++ semana-10/projeto/labex/src/assets/screens/CreateTripPage/CreateTripPage.js
@@ -15,7 +15,7 @@ const CreateTripPage = () => {
 
     // token ? getTripsList() : history.push("/login")
     token || history.push("/login")
-  })
+  }, [history])
 
   return (
     <div>
