+++ semana-10/projeto/labex/src/assets/screens/TripDetailsPage/TripDetailsPage.js
@@ -14,7 +14,7 @@ const TripDetailsPage = () => {
 
     // token ? getTripsList() : history.push("/login")
     token || history.push("/login")
-  })
+  }, [history])
 
   return (
     <div>
